engine-deals-ai/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
discord.py>=2.3.0
schedule>=1.2.0
webdriver-manager>=4.0.0
lxml>=4.9.0
fake-useragent>=1.4.0
joblib>=1.3.0
geopy>=2.3.0
